Complete GulpFile Setup Described in Getting Started with Gulp by Travis Maynard
This book provides a step-by-step overview of how to set up and use Gulp
as a task runner for CSS preprocessors, JavaScript minification,
LiveReload, and other tasks. This file includes integration of:

• beeper - a Node.js package that makes the Terminal beep. This is
configured to beep when an error occurs.
• browserify - a Node.js package that bundles JavaScript files and their
dependencies
• browser-sync - A Node.js package that provides LiveReload
functionality while synchronizing the reloads across multiple devices
connected to the same network. It allows you to live reload code changes on
desktops, tablets and phones at the same time.
• gulp-concat - a Gulp package that concatenates files. This is used in
the Scripts task to concatenate all JavaScript files included in the
project before minification
• del - a Node.js package that is used to delete files and folders using
globs. This is used in the Clean task which can be used to remove all
processed files from the dist\ directory before starting the task
runner. A 'reset' option.
• gulp-imagemin - a Gulp package that is used to minify images. This is
used in the Images task
• gulp-jshint - a Gulp package that detects errors and potential problems
in JavaScript files.
• gulp-plumber - a Gulp package that manages web asset pipelines. It is
used in the Styles task.
• gulp-sass - the Gulp package for my preferred CSS Preprocessor. The book introduces another
called gulp-myth which I replaced.
• gulp-sourcemaps - a Gulp package that can be used to create CSS and
JavaScript source maps which allow you to map code within a compressed
file back to it’s original position in a source file. This provides an
easy way to debug minified files which are otherwise unreadable.
• gulp-uglify - a Gulp package that allows you to use Gulp to minify
files using UglifyJS.
• vinyl-source-stream - Gulp uses a virtual file format
called vinyl to process files. With vinyl-source-stream, we can convert a file stream
to vinyl for further processing.